DEVEXP-350: Standardize UX for JS and XQY
This is a rather complex commit. It includes code refactoring, changes to functionality, and new functionality.

1) New functionality - The "Connected" status information from a previous commit only included information from XQY debug servers. This adds information from JS debug servers.

2) Changes to functionality - When connecting to a debug server previously, the user was presented with a pick-list for XQY debug servers, but had to type the app server name for JS debug servers. This has been changed so that it is always a pick-list.

3) Refactoring - The code to interact with XQY debug services and JS debug services is different. Additionally, the code to interact with those servers was in a required extension to a VSCode class (DebugConfigurationProvider). The refactoring splits that code into two separate classes (one for XQY and one for JS) to isolate it from the VSCode implementation and to make the code more consistent.
