+++ client/XQDebugger/xqyDebugConfigProvider.ts
@@ -3,31 +3,10 @@ import {
     window, DebugAdapterDescriptorFactory, DebugAdapterExecutable, DebugAdapterDescriptor,
     DebugSession, QuickPickItem, QuickPickOptions, WorkspaceConfiguration, workspace
 } from 'vscode';
-import { MarklogicClient, MlClientParameters, sendXQuery } from '../marklogicClient';
 import { readFileSync } from 'fs';
-import { MlxprsStatus } from '../mlxprsStatus';
+import { XqyDebugManager, DebugStatusQueryResponse } from './xqyDebugManager';
+import { MlClientParameters } from '../marklogicClient';
 
-const listServersQuery = `
-! (map:new()
-  => map:with("id", .)
-  => map:with("name", xdmp:server-name(.))
-  => map:with("port", xdmp:server-port(.)))
-=> xdmp:to-json()`;
-
-const listStoppedRequests = `
-for $rid in dbg:stopped()
-let $status := dbg:status($rid)/*:request
-let $server := xdmp:server-name($status/*:server-id)
-let $rstatus := $status/*:request-status/fn:string()
-let $dstatus := $status/*:debugging-status/fn:string()
-return (json:object()
-  => map:with('rid', $rid)
-  => map:with('serverName', $server)
-  => map:with('debugStatus', $dstatus)
-  => map:with('requestStatus', $rstatus)
-  )`;
-
-const doNotDebugThese = ['Admin', 'Manage', 'HealthCheck', 'App-Services'];
 
 export class XqyDebugConfiguration implements DebugConfiguration {
     [key: string]: any
@@ -49,36 +28,9 @@ const placeholder: QuickPickOptions = {
     placeHolder: 'Select the request to attach'
 };
 
-interface ServerQueryResponse {
-    name: string;
-    id: string;
-    port?: number;
-}
-
-interface DebugStatusQueryResponse {
-    rid: string;
-    serverName: string;
-    debugStatus: string;
-    requestStatus: string;
-}
-
 export class XqyDebugConfigurationProvider implements DebugConfigurationProvider {
-    static mlxprsStatus: MlxprsStatus = null;
 
-    private async getAvailableRequests(params: MlClientParameters): Promise<Array<DebugStatusQueryResponse>> {
-        const client: MarklogicClient = new MarklogicClient(params);
-        const resp = await sendXQuery(client, listStoppedRequests)
-            .result(
-                (fulfill: Record<string, any>[]) => {
-                    return fulfill.map(f => f.value as DebugStatusQueryResponse);
-                },
-                (error: Record<string, any>[]) => {
-                    console.error(JSON.stringify(error));
-                    return [];
-                });
-        return resp;
-    }
-
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     private async resolveRemainingDebugConfiguration(folder: WorkspaceFolder | undefined, config: XqyDebugConfiguration, token?: CancellationToken): Promise<DebugConfiguration> {
         const cfg: WorkspaceConfiguration = workspace.getConfiguration();
         const clientParams: MlClientParameters = new MlClientParameters({
@@ -97,7 +49,7 @@ export class XqyDebugConfigurationProvider implements DebugConfigurationProvider
         config.clientParams = clientParams;
 
         if (config.request === 'attach' && !config.rid) {
-            const rid: string = await this.getAvailableRequests(clientParams)
+            const rid: string = await XqyDebugManager.getAvailableRequests(clientParams)
                 .then((requests: Array<DebugStatusQueryResponse>) => {
                     if (requests.length) {
                         const qpRequests: QuickPickItem[] = requests.map((request: DebugStatusQueryResponse) => {
@@ -125,6 +77,7 @@ export class XqyDebugConfigurationProvider implements DebugConfigurationProvider
     /**
      * @override
      */
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     resolveDebugConfiguration(folder: WorkspaceFolder | undefined, config: XqyDebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration> {
         if (!config.type && !config.request && !config.name) {
             const editor = window.activeTextEditor;
@@ -145,60 +98,6 @@ export class XqyDebugConfigurationProvider implements DebugConfigurationProvider
         }
 
         return this.resolveRemainingDebugConfiguration(folder, config);
-    }
-
-
-    public static async chooseXqyServer(mlClient: MarklogicClient, intention: 'connect' | 'disconnect'): Promise<void> {
-        const command = intention === 'connect' ? 'xdmp:servers()' : 'dbg:connected()';
-        await sendXQuery(mlClient, `${command} ${listServersQuery}`)
-            .result(
-                (fulfill: Record<string, ServerQueryResponse>) => {
-                    const servers: ServerQueryResponse[] = [].concat(fulfill[0]['value'] || []);
-                    return servers
-                        .filter((server: ServerQueryResponse) => ((server.port || 0) !== mlClient.params.port || intention === 'disconnect'))
-                        .filter((server: ServerQueryResponse) => (!doNotDebugThese.includes(server.name) || intention === 'disconnect'))
-                        .map((server: ServerQueryResponse) => {
-                            return {
-                                label: server.name,
-                                description: server.id,
-                                detail: `${server.name} on ${mlClient.params.host}:${server.port || '(none)'}`,
-                            } as QuickPickItem;
-                        });
-                },
-                err => {
-                    window.showErrorMessage(`couldn't get a list of servers: ${JSON.stringify(err)}`);
-                    return [];
-                })
-            .then((choices: QuickPickItem[]) => {
-                if (choices.length) {
-                    return window.showQuickPick(choices)
-                        .then((choice: QuickPickItem) => {
-                            return sendXQuery(mlClient, `dbg:${intention}(${choice.description})`)
-                                .result(
-                                    () => {
-                                        window.showInformationMessage(`Successfully ${intention}ed ${choice.label} on ${mlClient.params.host}`);
-                                        this.requestStatusBarItemUpdate();
-                                    },
-                                    (err) => {
-                                        window.showErrorMessage(`Failed to connect to ${choice.label}: ${JSON.stringify(err.body.errorResponse.message)}`);
-                                        this.requestStatusBarItemUpdate();
-                                    });
-                        });
-                }
-                window.showWarningMessage(`No stopped servers found on ${mlClient.params.host}`);
-                this.requestStatusBarItemUpdate();
-                return null;
-            });
-    }
-
-    private static requestStatusBarItemUpdate() {
-        if (XqyDebugConfigurationProvider.mlxprsStatus) {
-            XqyDebugConfigurationProvider.mlxprsStatus.requestUpdate();
-        }
-    }
-
-    public static registerMlxprsStatusBarItem(mlxprsStatus: MlxprsStatus) {
-        XqyDebugConfigurationProvider.mlxprsStatus = mlxprsStatus;
     }
 }
