+++ client/marklogicClient.ts
@@ -1,5 +1,6 @@
 'use strict';
 
+import * as request from 'request-promise';
 import * as ml from 'marklogic';
 import * as fs from 'fs';
 import { MlxprsStatus } from './mlxprsStatus';
@@ -7,6 +8,12 @@ import { MlxprsStatus } from './mlxprsStatus';
 export const MLDBCLIENT = 'mldbClient';
 export const MLSETTINGSFLAG = /mlxprs:settings/;
 export const XQY = 'xqy';
+
+export interface ServerQueryResponse {
+    name: string;
+    id: string;
+    port?: number;
+}
 
 export class MlClientParameters {
     contentDb: string;
@@ -73,6 +80,7 @@ export class MlClientParameters {
 }
 
 export class MarklogicClient {
+    public static DEFAULT_MANAGE_PORT = 8002;
     params: MlClientParameters;
     ca: string;
 
@@ -109,21 +117,90 @@ export class MarklogicClient {
     }
 
 
-    async getConnectedServers(requestingObject: MlxprsStatus, updateCallback: (connectedServers: string[]) => void): Promise<void> {
+    async getConnectedServers(requestingObject: MlxprsStatus, managePort: number, updateCallback: (connectedServers: string[]) => void): Promise<void> {
+        const connectedJsServers: MlxprsQuickPickItem[] = await getFilteredListOfJsAppServers(this, managePort, 'true');
         return this.mldbClient.xqueryEval('xquery version "1.0-ml"; dbg:connected()')
             .result(
-                (items: ml.Item[]) => {
+                async (items: ml.Item[]) => {
                     const connectedServers: string[] = [];
-                    items.forEach((item) => {
-                        connectedServers.push(item.value);
+                    const requests = [];
+                    items.forEach(async (item) => {
+                        requests.push(
+                            sendXQuery(this, `xdmp:server-name(${item.value})`)
+                                .result(
+                                    (result: ml.Item[]) => {
+                                        connectedServers.push('XQY:' + result[0].value);
+                                    },
+                                    err => {
+                                        return [];
+                                    })
+                        );
                     });
-                    // requestingObject['updateStatusBarItem'](connectedServers);
-                    // requestingObject[updateCallback](connectedServers);
+                    await Promise.all(requests);
+
+                    connectedJsServers.forEach((quickPickItem) => {
+                        connectedServers.push('JS:' + quickPickItem.label);
+                    });
                     updateCallback.call(requestingObject, connectedServers);
                 },
                 (err) => {
                     throw err;
-                });
+                }
+            );
+    }
+
+    public listServersForDisconnectQuery = `
+    let $connected-app-servers := dbg:connected()
+    let $json-servers := $connected-app-servers ! (map:new()
+      => map:with("id", .)
+      => map:with("name", xdmp:server-name(.))
+      => map:with("port", xdmp:server-port(.)))
+    => xdmp:to-json()
+    return $json-servers
+    `;
+
+    public buildListServersForConnectQuery() {
+        return `
+let $all-app-servers := xdmp:servers()
+let $connected-app-servers := dbg:connected()
+let $not-connected-app-servers :=
+  for $app-server in $all-app-servers
+  return
+    if ($app-server = $connected-app-servers) then
+      ()
+    else
+      if (xdmp:server-name($app-server) = ('Admin', 'Manage', 'HealthCheck', 'App-Services')) then
+        ()
+      else
+        if (xdmp:server-port($app-server) = ${this.params.port}) then
+            ()
+        else
+            $app-server
+let $json-servers := $not-connected-app-servers ! (map:new()
+  => map:with("id", .)
+  => map:with("name", xdmp:server-name(.))
+  => map:with("port", xdmp:server-port(.)))
+=> xdmp:to-json()
+return $json-servers
+`;
+    }
+
+    public buildAppServerChoicesFromServerList(appServers: Record<string, ServerQueryResponse>) {
+        const servers: ServerQueryResponse[] = [].concat(appServers[0]['value'] || []);
+        return servers
+            .map((server: ServerQueryResponse) => {
+                return {
+                    label: server.name,
+                    description: server.id,
+                    detail: `${server.name} on ${this.params.host}:${server.port || '(none)'}`,
+                } as any;
+            });
+    }
+
+    public static buildUrl(hostname: string, endpointPath: string, ssl = true, managePort = MarklogicClient.DEFAULT_MANAGE_PORT): string {
+        const scheme: string = ssl ? 'https' : 'http';
+        const url = `${scheme}://${hostname}:${managePort}${endpointPath}`;
+        return url;
     }
 
 }
@@ -247,4 +324,65 @@ function sendRowsSerialized(db: MarklogicClient, actualQuery: string, resultForm
             }
         );
     }
+}
+
+export interface MlxprsQuickPickItem {
+    label: string;
+    description: string;
+    detail: string;
+}
+
+export async function getFilteredListOfJsAppServers(
+    mlClient: MarklogicClient, managePort: number, requiredResponse: string
+): Promise<MlxprsQuickPickItem[]> {
+    const listServersForConnectQuery = mlClient.buildListServersForConnectQuery();
+    const allPossibleJsServers: MlxprsQuickPickItem[] = await getAppServerListForJs(mlClient, listServersForConnectQuery) as MlxprsQuickPickItem[];
+    return await filterJsServerByConnectedStatus(mlClient, managePort, allPossibleJsServers, requiredResponse);
+}
+
+export async function getAppServerListForJs(mlClient: MarklogicClient, serverListQuery: string): Promise<MlxprsQuickPickItem[]> {
+    return sendXQuery(mlClient, serverListQuery)
+        .result(
+            (appServers: Record<string, ServerQueryResponse>) => {
+                return mlClient.buildAppServerChoicesFromServerList(appServers);
+            },
+            err => {
+                return [];
+            });
+}
+
+export async function filterJsServerByConnectedStatus(
+    mlClient: MarklogicClient, managePort: number, choices: MlxprsQuickPickItem[], requiredResponse: string
+): Promise<MlxprsQuickPickItem[]> {
+    const requests = [];
+    const filteredChoices: MlxprsQuickPickItem[] = [];
+    choices.forEach(async (choice) => {
+        const url = MarklogicClient.buildUrl(mlClient.params.host, `/jsdbg/v1/connected/${choice.label}`, mlClient.params.ssl, managePort);
+        const options = {
+            headers: {
+                'Content-type': 'application/x-www-form-urlencoded',
+                'X-Error-Accept': 'application/json'
+            },
+            auth: {
+                user: mlClient.params.user,
+                pass: mlClient.params.pwd,
+                'sendImmediately': false
+            }
+        };
+        if (mlClient.params.pathToCa !== '') {
+            options['agentOptions'] = { ca: fs.readFileSync(mlClient.params.pathToCa) };
+        }
+        options['rejectUnauthorized'] = mlClient.params.rejectUnauthorized;
+        const connectedRequest = request.get(url, options)
+            .then((response) => {
+                if (response === requiredResponse) {
+                    filteredChoices.push(choice);
+                }
+            }).catch(err => {
+                console.debug(`"Connected" request failed: ${err}`);
+            });
+        requests.push(connectedRequest);
+    });
+    await Promise.all(requests);
+    return filteredChoices;
 }
